+++ djangocodemirror/__init__.py
@@ -2,4 +2,4 @@
 """
 Django CodeMirror is a Django application to embed CodeMirror
 """
-__version__ = '1.0.2'
+__version__ = '1.0.3'